requirements.txt
flask
configobj
requests
python-telegram-bot